Inline basis worker to get around CSP/CDN issues
